+++ packages/web/src/javascripts/Components/ContentListView/Daily/CreateDailySection.spec.ts
@@ -1,0 +1,9 @@
+import { createDailyItemsWithToday } from './CreateDailySections'
+
+describe('create daily sections', () => {
+  it('createDailyItemsWithToday', () => {
+    const result = createDailyItemsWithToday(10)
+
+    expect(result).toHaveLength(10)
+  })
+})