+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.186](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.185...@standardnotes/releases@1.3.186) (2022-10-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.185](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.184...@standardnotes/releases@1.3.185) (2022-10-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.184](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.183...@standardnotes/releases@1.3.184) (2022-10-28)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.183](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.182...@standardnotes/releases@1.3.183) (2022-10-27)
 
 **Note:** Version bump only for package @standardnotes/releases